backend/pyproject.toml
[tool.poetry]
name = "ResoBin"
version = "0.4.0"
description = "Course resources by the insti for the insti"
authors = ["Laxman Desai <lolopaw24@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
django-heroku = "^0.3.1"
flake8 = "^3.9.2"
python-decouple = "^3.4"
black = "^21.6b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
